[GLIMMER] Cleanup Glimmer View Support Mixin
In glimmer we have a better strategy around setting layout and thus
private APIS such as `templateNameFor` are no longer needed. This also
removes the following

- CP for template (does nothing today)
- CP for layout (cannot change post init, slotted in component class)
- Slotted layoutName (in Component class)
- Slotted templateName (doesn't work in components and was deprecated)
- Slotted isView (not used in glimmer)
