+++ lib/packages.js
@@ -42,7 +42,7 @@ var glimmerStatus = features['ember-glimmer'];
 if (glimmerStatus === null || glimmerStatus === true) {
   packages['ember-glimmer'] = {
     trees: null,
-    requirements: ['container', 'ember-metal', 'ember-routing'],
+    requirements: ['container', 'ember-metal', 'ember-routing', 'ember-templates'],
     hasTemplates: true,
     vendorRequirements: [
       'glimmer',
