+++ packages/ember-templates/lib/utils/lookup-component.js
